# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для генерации и проверки TOTP-кодов, отправляемых по SMS",
        "title": "TOTP SMS Auth API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Сгенерировать и отправить TOTP-код по номеру телефона",
                "parameters": [
                    {
                        "description": "Phone",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Код отправлен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Проверить введённый TOTP-код",
                "parameters": [
                    {
                        "description": "Проверка кода",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.VerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Код верен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Неверный код",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.SendRequest": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string"
                }
            }
        },
        "main.VerifyRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    }
}